leptos_flow_core::groups

Struct GroupManager

pub struct GroupManager { /* private fields */ }

Manager for all group operations

Implementations

impl GroupManager

pub fn new() -> Self

Create a new group manager

pub fn create_group( &mut self, group_id: GroupId, members: HashSet<NodeId>, ) -> Result<()>

Create a group from a set of nodes

pub fn get_group(&self, group_id: &GroupId) -> Option<&Group>

Get a group by ID

pub fn get_group_mut(&mut self, group_id: &GroupId) -> Option<&mut Group>

Get a mutable group by ID

pub fn get_node_group(&self, node_id: &NodeId) -> Option<&GroupId>

Find which group contains a node

pub fn dissolve_group(&mut self, group_id: &GroupId) -> Result<HashSet<NodeId>>

Dissolve a group, removing all memberships

pub fn all_groups(&self) -> &HashMap<GroupId, Group>

Get all groups

pub fn node_mappings(&self) -> &HashMap<NodeId, GroupId>

Get all node-to-group mappings

pub fn add_node_to_group( &mut self, group_id: &GroupId, node_id: NodeId, ) -> Result<()>

Add a node to an existing group

pub fn remove_node_from_group( &mut self, group_id: &GroupId, node_id: &NodeId, ) -> Result<()>

Remove a node from a group

pub fn set_group_name( &mut self, group_id: &GroupId, name: Option<String>, ) -> Result<()>

Set the name of a group

pub fn calculate_group_bounds<N, E>( &mut self, group_id: &GroupId, graph: &Graph<N, E>, ) -> Result<()>
where N: Clone, E: Clone,

Calculate the bounding rectangle for a group based on its member positions

pub fn move_group<N, E>( &self, group_id: &GroupId, delta: Position, graph: &mut Graph<N, E>, ) -> Result<()>
where N: Clone, E: Clone,

Move all nodes in a group by a delta offset

pub fn is_point_in_group(&self, group_id: &GroupId, point: Position) -> bool

Check if a position is within the bounds of a group

pub fn start_group_drag<N, E>( &mut self, group_id: &GroupId, start_position: Position, graph: &Graph<N, E>, ) -> Result<()>
where N: Clone, E: Clone,

Start dragging a group

pub fn update_group_drag<N, E>( &mut self, new_position: Position, graph: &mut Graph<N, E>, ) -> Result<Position>
where N: Clone, E: Clone,

Update group drag position and move all nodes

pub fn complete_group_drag(&mut self) -> Result<GroupId>

Complete group drag operation

pub fn cancel_group_drag<N, E>( &mut self, graph: &mut Graph<N, E>, ) -> Result<GroupId>
where N: Clone, E: Clone,

Cancel group drag operation and restore original positions

pub fn is_group_dragging(&self) -> bool

Check if a group is currently being dragged

pub fn get_dragging_group(&self) -> Option<&GroupId>

Get the current dragging group ID (if any)

pub fn get_drag_delta(&self) -> Option<Position>

Get the current drag delta (if any)